test(memory): move the carve-out tests to a sibling file and round-trip them
Review follow-ups on #141:

- The tests were an inline `mod` in `mod.rs`; every other test module in this
  directory is a `#[path = "*_tests.rs"]` sibling. Now they match.
- The serde test only serialised. These types exist to cross a boundary, so a
  `Deserialize` derive that got gated away would not have failed it — it now
  round-trips and asserts the restored fields.
